Squirtvana PWA Features and Mobile OBS Integration (2)/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
openai==1.54.4
elevenlabs==2.6.0
websocket-client==1.8.0
requests==2.32.3
python-telegram-bot==21.7
psutil==6.1.0
python-dotenv==1.0.1

